chore: restructure components and update file paths style: update footer hover effect and add styles to center section
- Deleted:
  - src/components/ArticleHeading/ArticleHeading.jsx
  - src/components/ArticleHeading/articleHeading.css
  - src/components/ArticleText/ArticleText.jsx
  - src/components/ArticleText/articleText.css
  - src/components/Buttons/Button.jsx
  - src/components/Buttons/button.css
  - src/components/PrimaryArticleHeading/PrimaryArticleHeading/PrimaryArticleHeading.jsx
  - src/components/PrimaryArticleHeading/PrimaryArticleHeading/primaryArticleHeading.css

- Modified:
  - src/layout/footer/Footer.jsx
  - src/layout/footer/footer.css (added hover effect)
  - src/pages/home/Home.jsx
  - src/pages/home/centerSection/CenterSection.jsx (added CSS code)
  - src/pages/home/centerSection/centerSection.css
  - src/pages/home/sideSection/SideSection.jsx
  - src/pages/home/sideSection/sideSection.css

- Added:
  - src/components/ArticleHeading.jsx
  - src/components/ArticleText.jsx
  - src/components/Button.jsx
  - src/components/PrimaryArticleHeading.jsx
