--- src/components/ArticleHeading/articleHeading.css
@@ -1,6 +1,0 @@
-.new-content__article-header {
-    color: var(--color-almost-white);
-}
-.new-content__article-header:hover{
-    color: var(--color-light-vermillion);
-}
