+++ src/components/ArticleText.jsx
@@ -1,7 +1,5 @@
 import React from 'react'
 
-import './articleText.css'
-
 function ArticleText({ className, children }) {
     return <p className={className}>{children}</p>;
   }
