+++ src/components/Button.jsx
@@ -1,6 +1,4 @@
 
-import './button.css';
-
 function Button({className , children}) {
   return (
     <button className={className}>{children}</button>
