--- src/components/Buttons/button.css
@@ -1,15 +1,0 @@
-.button {
-    background-color: var(--color-light-vermillion);
-    border: none;
-    height: 4.8rem;
-    width: 18.5rem;
-    color: white;
-    font-weight: bold;
-    letter-spacing: 0.48rem;
-    text-transform: uppercase;
-    cursor: pointer;
-}
-
-.button:hover {
-    background-color: var(--color-dark-space-blue);
-}
