+++ src/components/PrimaryArticleHeading.jsx
@@ -1,6 +1,4 @@
 
-import './primaryArticleHeading.css'
-
 function PrimaryArticleHeading({className, children}) {
    return (
     <h2 className={className}>{children}</h2>
