+++ src/pages/home/centerSection/centerSection.css
@@ -1,41 +1,73 @@
 .center-section {
-    display:grid;
-    grid-template-rows: repeat(2, min-content);
+  display: grid;
+  grid-template-rows: repeat(2, min-content);
+  gap: 3rem;
+
+  .primary-image {
+    min-height: 30rem;
+    background-image: url(../../../assets/images/center-layout-images/image-web-3-desktop.jpg);
+  }
+
+  .primary-article {
+    display: flex;
     gap: 3rem;
 
-    .primary-image{
-        min-height: 30rem;
-        background-image:url(../../../assets/images/center-layout-images/image-web-3-desktop.jpg);
-        }
-
-    .primary-article{
-        display: flex;
-        gap: 3rem;
-
-        .primary-article__heading, .primary-article__description {
-            width: 50%;
-        }
-
-        .primary-article__description {
-            height: 100%;
-            display: flex;
-            flex-direction: column;
-            justify-content: space-between;
-            padding-right: 1.8rem;
-        }
+    .primary-article__heading,
+    .primary-article__description {
+      width: 50%;
     }
-}
 
-@media (max-width:60rem) {
+    .primary-article__heading {
+      color: var(--color-dark-space-blue);
 
-    .primary-article{
-        display: flex;
-        flex-direction: column;
-        gap: 1.6rem;
+      font-size: 5.6rem;
+      font-weight: bolder;
+      line-height: 5.6rem;
+      padding-right: 5.7rem;
+    }
 
     .primary-article__description {
-            width: 100% !important;
-        }
+      height: 100%;
+      display: flex;
+      flex-direction: column;
+      justify-content: space-between;
+      padding-right: 1.8rem;
 
+      .button {
+        background-color: var(--color-light-vermillion);
+        border: none;
+        height: 4.8rem;
+        width: 18.5rem;
+        color: white;
+        font-weight: bold;
+        letter-spacing: 0.48rem;
+        text-transform: uppercase;
+        cursor: pointer;
+      }
+
+      .button:hover {
+        background-color: var(--color-dark-space-blue);
+      }
     }
+  }
+}
+
+@media (max-width: 60rem) {
+  .primary-article {
+    display: flex;
+    flex-direction: column;
+    gap: 1.6rem;
+
+    .primary-article__description {
+      width: 100% !important;
+    }
+
+    .primary-article__heading {
+      width: 100% !important;
+    }
+
+    .primary-article__description {
+      gap: 2.4rem;
+    }
+  }
 }
