+++ src/pages/home/sideSection/sideSection.css
@@ -18,5 +18,15 @@
         display: flex;
         flex-direction: column;
         justify-content: space-between;
+
+        .new-content__article {
+
+            .new-content__article-header {
+                color: var(--color-almost-white);
+            }
+            .new-content__article-header:hover{
+                color: var(--color-light-vermillion);
+            }
+        }
     }
 }
